Restructure CLI into local/ and cloud/ modules
Moves local command code into src/local/ (cli.rs, mod.rs, output.rs,
server.rs) and cloud CLI types into src/cloud/cli.rs. Root cli.rs is
now a thin shell (~136 lines) with just Cli, Commands, SkillsArgs, and
re-exports. Local changes no longer require touching shared files.
