+++ src/cloud/mod.rs
@@ -1,4 +1,5 @@
 pub mod auth;
+pub mod cli;
 pub mod client;
 pub mod commands;
 pub mod credentials;
